refactor: delete -> deleteById로 네이밍 통일
