+++ src/main/java/nextstep/auth/controller/AuthController.java
@@ -1,10 +1,10 @@
-package nextstep.auth;
+package nextstep.auth.controller;
 
 
-import javax.servlet.http.HttpServletRequest;
-import nextstep.member.MemberResponse;
+import nextstep.auth.AuthService;
+import nextstep.auth.TokenRequest;
+import nextstep.auth.TokenResponse;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
